requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
Pillow==10.0.1
gunicorn==21.2.0
Werkzeug==2.3.7
Flask-CORS==4.0.0